refactor: clean up code
