+++ daytistics/lib/ui/onboarding/views/onboarding_view.dart
@@ -1,8 +1,7 @@
+import 'package:daytistics/application/providers/services/notification/notification_service.dart';
 import 'package:daytistics/application/providers/services/onboarding/onboarding_service.dart';
 import 'package:daytistics/application/providers/services/settings/settings_service.dart';
 import 'package:daytistics/config/settings.dart';
-import 'package:daytistics/main.dart';
-import 'package:daytistics/notifications.dart';
 import 'package:daytistics/shared/presets/home_view_preset.dart';
 import 'package:daytistics/shared/utils/dialogs.dart';
 import 'package:daytistics/shared/utils/internet.dart';
@@ -199,11 +198,12 @@ class _OnboardingViewState extends ConsumerState<OnboardingView> {
             ),
             TextButton(
               onPressed: () async {
-                maybeAskAllowNotifications();
                 await ref
                     .read(settingsServiceProvider)
                     .updateDailyReminderTime(timeOfDay: timeOfDay);
-                await scheduleDailyReminderNotification(timeOfDay);
+                await ref
+                    .read(notificationServiceProvider)
+                    .scheduleDailyReminderNotification(timeOfDay);
                 if (context.mounted) Navigator.of(context).pop();
                 onDone();
               },
